Rename solvable to variable
