+++ gaphas/solver/__init__.py
@@ -1,5 +1,5 @@
 from gaphas.solver.constraint import Constraint, MultiConstraint
-from gaphas.solver.solver import JuggleError, Solver, solvable
+from gaphas.solver.solver import JuggleError, Solver
 from gaphas.solver.variable import (
     NORMAL,
     REQUIRED,
@@ -8,4 +8,5 @@ from gaphas.solver.variable import (
     VERY_WEAK,
     WEAK,
     Variable,
+    variable,
 )
